sp-daemon: close all filedescriptors when we are doing a fork
Otherwise we are leaking the filedescriptors into the processes we are
later forking off.

The theory here is that all file descriptors that are existing longer
than one function call are registered in the manager, so they get
called. Once we are forking we are just closing all those
filedescriptors in the manager. and other should not exists or should be
closed by that time.